Restore Controller Manager package.
