+++ controller_manager/test/test_controller_manager_srvs.cpp
@@ -18,17 +18,72 @@
 #include <string>
 #include <vector>
 
-#include "controller_manager_test_common.hpp"
-
 #include "controller_interface/controller_interface.hpp"
 #include "controller_manager/controller_manager.hpp"
 #include "controller_manager_msgs/srv/list_controller_types.hpp"
 #include "controller_manager_msgs/srv/list_controllers.hpp"
 #include "controller_manager_msgs/srv/switch_controller.hpp"
+#include "controller_manager_test_common.hpp"
 #include "lifecycle_msgs/msg/state.hpp"
 
 using ::testing::_;
 using ::testing::Return;
+
+using namespace std::chrono_literals;
+
+class TestControllerManagerSrvs : public ControllerManagerFixture
+{
+public:
+  TestControllerManagerSrvs() {}
+
+  void SetUp() override
+  {
+    ControllerManagerFixture::SetUp();
+
+    update_timer_ = cm_->create_wall_timer(std::chrono::milliseconds(10), [&]() {
+      cm_->read();
+      cm_->update(rclcpp::Time(0), rclcpp::Duration::from_seconds(0.01));
+      cm_->write();
+    });
+
+    executor_->add_node(cm_);
+
+    executor_spin_future_ = std::async(std::launch::async, [this]() -> void { executor_->spin(); });
+    // This sleep is needed to prevent a too fast test from ending before the
+    // executor has began to spin, which causes it to hang
+    std::this_thread::sleep_for(50ms);
+  }
+
+  void TearDown() override { executor_->cancel(); }
+
+  template <typename T>
+  std::shared_ptr<typename T::Response> call_service_and_wait(
+    rclcpp::Client<T> & client, std::shared_ptr<typename T::Request> request,
+    rclcpp::Executor & service_executor, bool update_controller_while_spinning = false)
+  {
+    EXPECT_TRUE(client.wait_for_service(std::chrono::milliseconds(500)));
+    auto result = client.async_send_request(request);
+    // Wait for the result.
+    if (update_controller_while_spinning)
+    {
+      while (service_executor.spin_until_future_complete(result, 50ms) !=
+             rclcpp::FutureReturnCode::SUCCESS)
+      {
+        cm_->update(rclcpp::Time(0), rclcpp::Duration::from_seconds(0.01));
+      }
+    }
+    else
+    {
+      EXPECT_EQ(
+        service_executor.spin_until_future_complete(result), rclcpp::FutureReturnCode::SUCCESS);
+    }
+    return result.get();
+  }
+
+protected:
+  rclcpp::TimerBase::SharedPtr update_timer_;
+  std::future<void> executor_spin_future_;
+};
 
 TEST_F(TestControllerManagerSrvs, list_controller_types)
 {
@@ -146,20 +201,16 @@ TEST_F(TestControllerManagerSrvs, list_controllers_srv)
   ASSERT_EQ("active", result->controller[0].state);
   ASSERT_THAT(
     result->controller[0].claimed_interfaces,
-    testing::ElementsAre(
-      "joint2/max_acceleration", "configuration/max_tcp_jerk", "joint2/velocity", "joint3/velocity",
-      "joint1/max_velocity", "joint1/position"));
+    testing::ElementsAre("joint1/position", "joint2/velocity", "joint3/velocity"));
   ASSERT_THAT(
     result->controller[0].required_command_interfaces,
-    testing::ElementsAre(
-      "configuration/max_tcp_jerk", "joint1/max_velocity", "joint1/position",
-      "joint2/max_acceleration", "joint2/velocity", "joint3/velocity"));
+    testing::ElementsAre("joint1/position", "joint2/velocity", "joint3/velocity"));
   ASSERT_THAT(
     result->controller[0].required_state_interfaces,
     testing::ElementsAre(
-      "configuration/max_tcp_jerk", "joint1/position", "joint1/some_unlisted_interface",
-      "joint1/velocity", "joint2/acceleration", "joint2/position", "joint2/velocity",
-      "joint3/acceleration", "joint3/position", "joint3/velocity", "sensor1/velocity"));
+      "joint1/position", "joint1/some_unlisted_interface", "joint1/velocity", "joint2/acceleration",
+      "joint2/position", "joint2/velocity", "joint3/acceleration", "joint3/position",
+      "joint3/velocity", "sensor1/velocity"));
 
   cm_->switch_controller(
     {}, {test_controller::TEST_CONTROLLER_NAME},
@@ -171,15 +222,13 @@ TEST_F(TestControllerManagerSrvs, list_controllers_srv)
   ASSERT_TRUE(result->controller[0].claimed_interfaces.empty());
   ASSERT_THAT(
     result->controller[0].required_command_interfaces,
-    testing::ElementsAre(
-      "configuration/max_tcp_jerk", "joint1/max_velocity", "joint1/position",
-      "joint2/max_acceleration", "joint2/velocity", "joint3/velocity"));
+    testing::ElementsAre("joint1/position", "joint2/velocity", "joint3/velocity"));
   ASSERT_THAT(
     result->controller[0].required_state_interfaces,
     testing::ElementsAre(
-      "configuration/max_tcp_jerk", "joint1/position", "joint1/some_unlisted_interface",
-      "joint1/velocity", "joint2/acceleration", "joint2/position", "joint2/velocity",
-      "joint3/acceleration", "joint3/position", "joint3/velocity", "sensor1/velocity"));
+      "joint1/position", "joint1/some_unlisted_interface", "joint1/velocity", "joint2/acceleration",
+      "joint2/position", "joint2/velocity", "joint3/acceleration", "joint3/position",
+      "joint3/velocity", "sensor1/velocity"));
 
   ASSERT_EQ(
     controller_interface::return_type::OK,
