+++ hardware_interface/doc/hardware_components_userdoc.rst
@@ -7,49 +7,6 @@ There are three types of hardware Actuator, Sensor and System.
 For details on each type check `Hardware Components description <https://ros-controls.github.io/control.ros.org/getting_started.html#hardware-components>`_.
 
 
-Lifecycle of Hardware Components
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-Definitions and Nomenclature
-,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
-
-Hardware interfaces use the lifecycle state machine `defined for ROS2 nodes <https://design.ros2.org/articles/node_lifecycle.html>`_.
-There is only one addition to the state machine, that is the initialization method providing hardware configuration from URDF file as argument.
-
-Hardware Interface
-  User impl....
-
-Hardware Components
-  Wrapper and abstraction of hardware interface to manage life cycle and access to methods of hardware interface from Resource Manager.
-
-Resource Manager
-  Class responsible for the management of hardware components in the ros2_control framework.
-
-"movement" command interfaces
-  Interfaces responsible for robot to move, i.e., influence its dynamic behavior.
-  The interfaces are defined in `hardware_interface_type_values.hpp <https://github.com/ros-controls/ros2_control/blob/master/hardware_interface/include/hardware_interface/types/hardware_interface_type_values.hpp>`_. (TODO: add link to doxygen)
-
-"non-movement" command interfaces
-  All other interfaces that are not "movement" command interfaces (TODO: add link to def.)
-
-
-Initialization
-,,,,,,,,,,,,,,,
-Immediately after a plugin in loaded and the object created with the default constructor, the ``on_init`` method will be called providing hardware URDF configuration using the ``HardwareInfo`` structure.
-In this stage you should initialize all memory you need and prepare storage for interfaces.
-The resource manager will export of all interfaces after this and store them internally.
-
-
-Configuration
-,,,,,,,,,,,,,,
-Precondition is hardware interface state having id: ``lifecycle_msgs::msg::State::PRIMARY_STATE_UNCONFIGURED``.
-After configuration, the ``read`` and ``write`` methods will be called in the update loop.
-This means all internal state and commands variables have to be initialized.
-After a successful call to ``on_configure``, all state interfaces and "non-movement" command interfaces should be available to controllers.
-
-NOTE: If using "non-movement" command interfaces to parametrize the robot in the ``lifecycle_msgs::msg::State::PRIMARY_STATE_CONFIGURED`` state make sure to take care about current state in the ``write`` method of your Hardware Interface implementation.
-
-
 Handling of errors that happen during read() and write() calls
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
